Move organizer dependencies from bootstrap to OrganizerServiceProvider
